feat: enhance PWA functionality with standalone visualizer mode
- Added support for standalone PWA mode in the Idling Radio app, ensuring it opens in fullscreen visualizer mode.
- Introduced a new component, RadioPwaStandaloneVisualizerBootstrap, to manage visualizer mode activation based on PWA status.
- Updated global CSS to accommodate standalone PWA layout, including new styles for visualizer mode.
- Refactored isStandalonePwa function for better modularity and reusability across components.
- Removed deprecated styles and components related to the previous layout structure.

These changes collectively improve the user experience for PWA users and streamline the visualizer functionality.
