+++ src/app/components/radio-pwa/radioPwaAccess.ts
@@ -1,22 +1,2 @@
 export { IDLING_RADIO_PWA_START_PATH as RADIO_PWA_START_PATH } from '@lib/radio-pwa/constants';
-
-export function isStandalonePwa(): boolean {
-  if (typeof window === 'undefined') {
-    return false;
-  }
-
-  if (window.matchMedia('(display-mode: standalone)').matches) {
-    return true;
-  }
-
-  const navigatorWithStandalone = window.navigator as Navigator & { standalone?: boolean };
-  if (navigatorWithStandalone.standalone === true) {
-    return true;
-  }
-
-  if (document.referrer.includes('android-app://')) {
-    return true;
-  }
-
-  return false;
-}
+export { isStandalonePwa } from '@lib/radio-pwa/isStandalonePwa';
